+++ crates/parserc/src/errors.rs
@@ -23,6 +23,10 @@ pub enum Kind {
     Keyword(ControlFlow, Span),
     #[error("Error from parsing syntax `{0}`")]
     Syntax(&'static str, ControlFlow, Span),
+    #[error("Error from parsing syntax `LimitsTo`")]
+    LimitsTo(ControlFlow, Span),
+    #[error("Error from parsing syntax `Limits`")]
+    Limits(ControlFlow, Span),
 }
 
 /// A error type returns by parser combinators.
@@ -43,6 +47,8 @@ impl ParseError for Kind {
             Kind::NextIf(control_flow, _) => *control_flow,
             Kind::Keyword(control_flow, _) => *control_flow,
             Kind::Syntax(_, control_flow, _) => *control_flow,
+            Kind::LimitsTo(control_flow, _) => *control_flow,
+            Kind::Limits(control_flow, _) => *control_flow,
         }
     }
 
@@ -52,6 +58,8 @@ impl ParseError for Kind {
             Kind::NextIf(_, span) => Kind::NextIf(ControlFlow::Fatal, span),
             Kind::Keyword(_, span) => Kind::Keyword(ControlFlow::Fatal, span),
             Kind::Syntax(name, _, span) => Kind::Syntax(name, ControlFlow::Fatal, span),
+            Kind::LimitsTo(_, span) => Kind::LimitsTo(ControlFlow::Fatal, span),
+            Kind::Limits(_, span) => Kind::Limits(ControlFlow::Fatal, span),
         }
     }
 
@@ -61,6 +69,8 @@ impl ParseError for Kind {
             Kind::NextIf(_, span) => span.clone(),
             Kind::Keyword(_, span) => span.clone(),
             Kind::Syntax(_, _, span) => span.clone(),
+            Kind::LimitsTo(_, span) => span.clone(),
+            Kind::Limits(_, span) => span.clone(),
         }
     }
 }