+++ CHANGELOG.md
@@ -9,6 +9,11 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for
 compatibility with GitHub comment style markdown rendering.
 -->
 
+## [0.10.4] - 2025-10-30
+
+- syntax: add `Limits*` syntax nodes.
+- syntax: add `to_span` function to the `Syntax` trait.
+
 ## [0.10.3] - 2025-10-29
 
 - Change `ParseError` api: add `span` function.